+++ WalletWasabi.Fluent/Styles/Styles.xaml
@@ -5,14 +5,13 @@
   <Styles.Resources>
     <sys:Double x:Key="IconElementThemeHeight">16</sys:Double>
     <sys:Double x:Key="SplitViewCollapseThreshold">750</sys:Double>
+    <sys:Double x:Key="SplitViewCompactPaneLength">46</sys:Double>
   </Styles.Resources>
-
   <Style Selector="SplitView">
-    <Setter Property="CompactPaneLength" Value="46" />
+    <Setter Property="CompactPaneLength" Value="{StaticResource SplitViewCompactPaneLength}" />
     <Setter Property="OpenPaneLength" Value="320" />
     <Setter Property="DisplayMode" Value="CompactInline" />
     <Setter Property="IsPaneOpen" Value="True" />
     <Setter Property="PaneBackground" Value="{x:Null}" />
-  </Style>
-  
+  </Style>  
 </Styles>
