+++ WalletWasabi.Fluent/ViewModels/AddWalletPageViewModel.cs
@@ -1,0 +1,12 @@
+using ReactiveUI;
+
+namespace WalletWasabi.Fluent.ViewModels
+{
+	public class AddWalletPageViewModel : NavBarItemViewModel
+	{
+		public AddWalletPageViewModel(IScreen screen) : base(screen)
+		{
+			Title = "Add Wallet";
+		}
+	}
+}
