+++ WalletWasabi.Fluent/ViewModels/HomePageViewModel.cs
@@ -1,0 +1,12 @@
+using ReactiveUI;
+
+namespace WalletWasabi.Fluent.ViewModels
+{
+	public class HomePageViewModel : NavBarItemViewModel
+	{
+		public HomePageViewModel(IScreen screen) : base(screen)
+		{
+			Title = "Home";
+		}
+	}
+}
