+++ WalletWasabi.Fluent/ViewModels/NavBarItemViewModel.cs
@@ -1,40 +1,8 @@
 using ReactiveUI;
 using System;
-using System.Windows.Input;
 
 namespace WalletWasabi.Fluent.ViewModels
 {
-	public class HomePageViewModel : NavBarItemViewModel
-	{
-		public HomePageViewModel(IScreen screen) : base(screen)
-		{
-			Title = "Home";
-		}
-	}
-
-	public class SettingsPageViewModel : NavBarItemViewModel
-	{
-		public SettingsPageViewModel(IScreen screen) : base(screen)
-		{
-			Title = "Settings";
-
-			NextCommand = ReactiveCommand.Create(() =>
-			{
-				screen.Router.Navigate.Execute(new HomePageViewModel(screen));
-			});
-		}
-
-		public ICommand NextCommand { get; }
-	}
-
-	public class AddWalletPageViewModel : NavBarItemViewModel
-	{
-		public AddWalletPageViewModel(IScreen screen) : base(screen)
-		{
-			Title = "Add Wallet";
-		}
-	}
-
 	public abstract class NavBarItemViewModel : ViewModelBase, IRoutableViewModel
 	{
 		private bool _isSelected;
@@ -46,7 +14,7 @@ namespace WalletWasabi.Fluent.ViewModels
 			HostScreen = screen;
 		}
 
-		public string UrlPathSegment { get; } = Guid.NewGuid().ToString().Substring(0, 5);
+		public string UrlPathSegment => Guid.NewGuid().ToString().Substring(0, 5);
 
 		public IScreen HostScreen { get; }
 
@@ -64,8 +32,8 @@ namespace WalletWasabi.Fluent.ViewModels
 
 		public bool IsSelected
 		{
-			get { return _isSelected; }
-			set { this.RaiseAndSetIfChanged(ref _isSelected, value); }
+			get => _isSelected;
+			set => this.RaiseAndSetIfChanged(ref _isSelected, value);
 		}
 	}
 }
