+++ WalletWasabi.Fluent/ViewModels/ViewModelBase.cs
@@ -7,6 +7,5 @@ namespace WalletWasabi.Fluent.ViewModels
 {
 	public class ViewModelBase : ReactiveObject
 	{
-
 	}
 }
