+++ WalletWasabi.Fluent/Views/MainView.axaml
@@ -9,7 +9,7 @@
              xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors;assembly=WalletWasabi.Fluent"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="WalletWasabi.Fluent.Views.MainView" TextBlock.Foreground="#A3A3A3">
-  <SplitView CompactPaneLength="46" IsPaneOpen="{Binding NavBar.IsOpen, Mode=TwoWay}">
+  <SplitView IsPaneOpen="{Binding NavBar.IsOpen, Mode=TwoWay}">
     <i:Interaction.Behaviors>
         <behaviors:SplitViewAutoBehavior ToggleAction="{Binding NavBar.ToggleAction, Mode=OneWayToSource}"
                                          CollapseOnClickAction="{Binding NavBar.CollapseOnClickAction, Mode=OneWayToSource}"
