+++ WalletWasabi.Fluent/Views/NavBar.axaml
@@ -47,7 +47,7 @@
       <Setter Property="Margin" Value="-16 0 0 0" />
     </Style>
     <Style Selector=":is(Button).NavButton">
-      <Setter Property="Width" Value="46" />
+      <Setter Property="Width" Value="{StaticResource SplitViewCompactPaneLength}" />
       <Setter Property="Padding" Value="12,9,12,12" />
     </Style>
     <Style Selector=":is(Button).NavButton:disabled > PathIcon">
